refactor(heroes): make tabs at Heroes page, move Heroes Radar to HeroesRadarTab
